requirements.txt
pyTelegramBotAPI==4.4.0
python-dotenv==0.19.2
loguru==0.6.0
sqlalchemy==1.4.41
requests==2.28.1
python-telegram-bot-calendar